+++ app/Jobs/TelegramSendJob.php
@@ -4,10 +4,11 @@ namespace App\Jobs;
 
 use App\Http\TelegramBot\Aliases;
 use App\Http\TelegramBot\Auth\Authentication;
+use App\Http\TelegramBot\Components\RecursionClass\MenuR;
+use App\Http\TelegramBot\Info\Alerts\InputAlert;
 use App\Http\TelegramBot\Services\ArgumentsService;
 use App\Http\TelegramBot\States\StateMake;
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
@@ -38,19 +39,28 @@ class TelegramSendJob implements ShouldQueue
     {
         $authentication = new Authentication();
         $user = $authentication->handle($this->update);
-
-        if ($this->action === 'send'){
-            $user->state()->delete();
-            $className = Aliases::getFullNameSpace($this->argumentsService->cl);
-            $createdClass = new $className($user, $this->update, $this->argumentsService);
-            $createdClass->sendCreate();
-        }elseif ($this->action === 'edit'){
-            $className = Aliases::getFullNameSpace($this->argumentsService->cl);
-            $createdClass = new $className($user, $this->update, $this->argumentsService);
-            $createdClass->handleCallbackQuery();
-        }elseif ($this->action === 'state'){
-            if ($state = $user->state){
-                (new StateMake($this->update, $user, $this->argumentsService, $state))->make();
+        if ($user->is_blocked){
+            $this->argumentsService->er = '10';
+            (new InputAlert($user, $this->update,
+                $this->argumentsService))->handleCallbackQuery();
+        }else{
+            if ($this->action === 'send'){
+                $user->state()->delete();
+                $className = Aliases::getFullNameSpace($this->argumentsService->cl);
+                $createdClass = new $className($user, $this->update, $this->argumentsService);
+                $createdClass->sendCreate();
+            }elseif ($this->action === 'edit'){
+//                $argumentsService = new ArgumentsService('');
+//                $menuR = new MenuR($user, $this->update, $argumentsService);
+//                $menuR->handleCallbackQuery();
+                if ($this->argumentsService->s) $user->state()->delete();
+                $className = Aliases::getFullNameSpace($this->argumentsService->cl);
+                $createdClass = new $className($user, $this->update, $this->argumentsService);
+                $createdClass->handleCallbackQuery();
+            }elseif ($this->action === 'state'){
+                if ($state = $user->state){
+                    (new StateMake($this->update, $user, $this->argumentsService, $state))->make();
+                }
             }
         }
     }
